add v_id to bagname; upload to hadoop

+++ regression_test.sh
@@ -13,23 +13,29 @@ cd /home/autowise/qyc/models
 
 git pull http://git.autowise.ai/autowise/models
 
-bash -c "
-docker run --rm --name regression_test\
--v ~:/home/autowise/ \
--v /home/autowise/qyc/models/aw_models:/opt/ros/kinetic/share/aw_models \
--v /home/autowise/data/regression_test/:/home/autowise/data/regression_test/ \
--w /home/autowise/ --ulimit core=-1 --security-opt seccomp=unconfined \
---network host --privileged=true \
-registry.autowise.ai/awcar:latest bash -c \"source /opt/ros/kinetic/setup.bash;\
-source ~/.autowise/setup.sh;\
-sudo apt-get install ros-kinetic-autowise-tools; \
-cd /home/autowise/jenkins_ws/workspace/aw_autowise_test_branch/log_based_simu/; \
-./install_dependencies.sh; \
-export kRobotTypeHejia18T=hejia; \
-python regression_parallelly.py --save --record\" "
-if [ $? -ne 0 ];then
-    echo "This means something failed"
-    exit -1
+container=`docker ps | grep regression_test | head -n 1 | cut -d ' ' -f1`
+
+if [[ ! -n ${container} ]]; 
+then 
+    bash -c "
+    docker run --rm --name regression_test\
+    -v ~:/home/autowise/ \
+    -v /usr/lib/jvm/java-8-oracle:/usr/lib/jvm/java-8-oracle \
+    -v /home/autowise/qyc/models/aw_models:/opt/ros/kinetic/share/aw_models \
+    -v /home/autowise/data/regression_test/:/home/autowise/data/regression_test/ \
+    -w /home/autowise/ --ulimit core=-1 --security-opt seccomp=unconfined \
+    --network host --privileged=true \
+    registry.autowise.ai/awcar:latest bash -c \"source /opt/ros/kinetic/setup.bash;\
+    source ~/.autowise/setup.sh;\
+    sudo apt-get install ros-kinetic-autowise-tools; \
+    cd /home/autowise/jenkins_ws/workspace/aw_autowise_test_branch/log_based_simu/; \
+    ./install_dependencies.sh; \
+    export kRobotTypeHejia18T=hejia; \
+    python regression_parallelly.py --save --record\" "
+    if [ $? -ne 0 ];then
+        echo "This means something failed"
+        exit -1
+    fi
 fi
 
 echo '=======================END============================='
